Refactored project, added basic logging, and fixed the '.' path moving error
