+++ logging.text
@@ -1,0 +1,5 @@
+18:00:34,478 root WARNING Will not do this Modi_voluptatem.txt
+18:00:34,479 root WARNING Will not do this Non_dolor.csv
+18:00:34,479 root WARNING Will not do this Labore_dolorem.xls
+18:00:34,480 root WARNING Will not do this Etincidunt_ipsum_dolor.csv
+18:00:34,480 root WARNING Will not do this Est_velit_modi.csv
